# bff/openapi/bff.yaml
openapi: 3.0.3
info:
  title: MARRAKEA BFF API
  description: |
    Backend-For-Frontend API for the Marrakea Next.js frontend.
    Aggregates Strapi (editorial) and Medusa (commerce) data into strict DTOs.
  version: 1.0.0
  contact:
    name: Marrakea Team
servers:
  - url: https://api.marrakea.com
    description: Production
  - url: http://localhost:4000
    description: Development
...
paths:
  /catalog/products:
    get:
      tags: [Catalog]
      summary: List paginated products
      operationId: listProducts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
        - name: search
          in: query
          schema:
            type: string
          description: Search in title/intro
        - name: gesture
          in: query
          schema:
            type: string
          description: Filter by gesture slug
        - name: territory
          in: query
          schema:
            type: string
          description: Filter by territory slug
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, name_asc, price_asc, price_desc]
          description: Sort order (price_asc/price_desc return 400 in v1)
      responses:
        '200':
          description: Paginated list of products
          headers:
            Cache-Control:
              schema:
                type: string
                example: "public, s-maxage=60, stale-while-revalidate=300"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductResponse'
        '400':
          description: Unsupported sort parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /catalog/products/{slug}:
    get:
      tags: [Catalog]
      summary: Get product detail by slug
      operationId: getProductBySlug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product detail
          headers:
            Cache-Control:
              schema:
                type: string
                example: "public, s-maxage=60, stale-while-revalidate=300"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetailDTO'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ImageDTO:
      type: object
      properties:
        url:
          type: string
        alt:
          type: string
        width:
          type: integer
        height:
          type: integer
        blurDataUrl:
          type: string
      required:
        - url
        - alt
    MoneyDTO:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
          enum: [EUR]
        formattedPrice:
          type: string
      required:
        - amount
        - currency
        - formattedPrice
    AvailabilityDTO:
      type: object
      properties:
        inStock:
          type: boolean
        purchaseMode:
          type: string
          enum: [unique, standard]
        label:
          type: string
      required:
        - inStock
        - purchaseMode
        - label
    GestureDTO:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
      required:
        - id
        - name
        - slug
    TerritoryDTO:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
      required:
        - id
        - name
        - slug
    ArtisanDTO:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        bio:
          type: string
        portrait:
          $ref: '#/components/schemas/ImageDTO'
        territory:
          $ref: '#/components/schemas/TerritoryDTO'
        workshopLocation:
          type: string
        specialty:
          type: string
        yearsExperience:
          type: string
        transmissionMode:
          type: string
        equipment:
          type: string
      required:
        - id
        - name
    ProductDTO:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        intro:
          type: string
        coverImage:
          $ref: '#/components/schemas/ImageDTO'
        price:
          $ref: '#/components/schemas/MoneyDTO'
        availability:
          $ref: '#/components/schemas/AvailabilityDTO'
        gesture:
          $ref: '#/components/schemas/GestureDTO'
        territory:
          $ref: '#/components/schemas/TerritoryDTO'
        defaultVariantId:
          type: string
      required:
        - id
        - slug
        - title
        - coverImage
        - price
        - availability
        - gesture
        - territory
    ProductDetailDTO:
      allOf:
        - $ref: '#/components/schemas/ProductDTO'
        - type: object
          properties:
            images:
              type: array
              items:
                $ref: '#/components/schemas/ImageDTO'
              description: Secondary images (excludes coverImage)
            description:
              type: string
              description: HTML content (sanitized)
            dimensions:
              type: string
            materials:
              type: array
              items:
                type: string
            artisan:
              $ref: '#/components/schemas/ArtisanDTO'
            acquisition:
              type: string
            referenceSheet:
              type: object
              additionalProperties:
                type: string
          required:
            - images
...
    PaginatedProductResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductDTO'
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
        - data
        - pagination
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error
        - code